+++ core/embed/rust/Cargo.toml
@@ -11,12 +11,14 @@ bitcoin_only = []
 model_tt = ["touch"]
 model_t1 = ["buttons"]
 model_tr = ["buttons"]
+micropython = []
+protobuf = ["micropython"]
 ui = []
 ui_debug = []
 buttons = []
 touch = []
 clippy = []
-test = ["cc", "glob"]
+test = ["cc", "glob", "micropython", "protobuf", "ui", "ui_debug"]
 
 [lib]
 crate-type = ["staticlib"]
