+++ core/embed/rust/src/error.rs
@@ -1,7 +1,11 @@
-use core::convert::{Infallible, TryInto};
+use core::{
+    convert::{Infallible, TryInto},
+    num::TryFromIntError,
+};
 
 use cstr_core::CStr;
 
+#[cfg(feature = "micropython")]
 use crate::micropython::{ffi, obj::Obj, qstr::Qstr};
 
 #[allow(clippy::enum_variant_names)] // We mimic the Python exception classnames here.
@@ -11,13 +15,19 @@ pub enum Error {
     OutOfRange,
     MissingKwargs,
     AllocationFailed,
+    #[cfg(feature = "micropython")]
     CaughtException(Obj),
+    #[cfg(feature = "micropython")]
     KeyError(Obj),
+    #[cfg(feature = "micropython")]
     AttributeError(Qstr),
+    #[cfg(feature = "micropython")]
     ValueError(&'static CStr),
+    #[cfg(feature = "micropython")]
     ValueErrorParam(&'static CStr, Obj),
 }
 
+#[cfg(feature = "micropython")]
 impl Error {
     /// Create an exception instance matching the error code. The result of this
     /// call should only be used to immediately raise the exception, because the
@@ -65,5 +75,11 @@ impl Error {
 impl From<Infallible> for Error {
     fn from(_: Infallible) -> Self {
         unreachable!()
+    }
+}
+
+impl From<TryFromIntError> for Error {
+    fn from(_: TryFromIntError) -> Self {
+        Self::OutOfRange
     }
 }
