+++ core/embed/rust/src/micropython/util.rs
@@ -81,16 +81,3 @@ pub unsafe fn try_with_args_and_kwargs_inline(
     };
     unsafe { try_or_raise(block) }
 }
-
-pub trait ResultExt {
-    fn assert_if_debugging_ui(self, message: &str);
-}
-
-impl<T, E> ResultExt for Result<T, E> {
-    fn assert_if_debugging_ui(self, #[allow(unused)] message: &str) {
-        #[cfg(feature = "ui_debug")]
-        if self.is_err() {
-            panic!("{}", message);
-        }
-    }
-}
