+++ core/embed/rust/src/trezorhal/mod.rs
@@ -5,3 +5,8 @@ pub mod display;
 mod ffi;
 pub mod random;
 pub mod slip39;
+
+#[cfg(not(feature = "micropython"))]
+pub mod time;
+#[cfg(feature = "micropython")]
+pub use crate::micropython::time;
