+++ core/embed/rust/src/ui/display.rs
@@ -1,5 +1,8 @@
 use super::constant;
-use crate::{micropython::time, time::Duration, trezorhal::display};
+use crate::{
+    time::Duration,
+    trezorhal::{display, time},
+};
 
 use super::geometry::{Offset, Point, Rect};
