+++ core/embed/rust/src/ui/model_t1/mod.rs
@@ -1,4 +1,6 @@
 pub mod component;
 pub mod constant;
-pub mod layout;
 pub mod theme;
+
+#[cfg(feature = "micropython")]
+pub mod layout;
