+++ core/embed/rust/src/ui/model_tt/mod.rs
@@ -1,5 +1,7 @@
 pub mod component;
 pub mod constant;
 pub mod event;
-pub mod layout;
 pub mod theme;
+
+#[cfg(feature = "micropython")]
+pub mod layout;
