+++ core/embed/rust/src/ui/util.rs
@@ -1,0 +1,12 @@
+pub trait ResultExt {
+    fn assert_if_debugging_ui(self, message: &str);
+}
+
+impl<T, E> ResultExt for Result<T, E> {
+    fn assert_if_debugging_ui(self, #[allow(unused)] message: &str) {
+        #[cfg(feature = "ui_debug")]
+        if self.is_err() {
+            panic!("{}", message);
+        }
+    }
+}
